Cleanup

+++ src/main/java/theartifex/ui/CyberneticCampfireOption.java
@@ -53,7 +53,7 @@ public class CyberneticCampfireOption extends AbstractCampfireOption {
 
     @Override
     public void update() {
-        float hackScale = (float) ReflectionHacks.getPrivate(this, AbstractCampfireOption.class, "scale");
+        float hackScale = ReflectionHacks.getPrivate(this, AbstractCampfireOption.class, "scale");
 
         if (this.hb.hovered) {
